[test-framework] Refactoring local application and agent (#8809)
Merging manually since test framework does not affect production code and the failing test is unrelated. The test failure is likely due to something like https://github.com/istio/istio/pull/8052?.

+++ pkg/test/util/reserveport/api.go
@@ -21,7 +21,7 @@ const (
 // ReservedPort a port reserved by a PortManager
 type ReservedPort interface {
 	// GetPort returns the bound port number.
-	GetPort() uint32
+	GetPort() uint16
 	// Close unbinds this port.
 	Close() error
 }
